adjust tooth logger api
