+++ firmware/tunerstudio/rusefi.input
@@ -3565,6 +3565,7 @@ cmd_set_engine_type_default					= "@@TS_IO_TEST_COMMAND_char@@@@ts_command_e_TS_
 		field = "CS Pin",								sdCardCsPin			@@if_ts_show_sd_pins
 		field = "SPI",									sdCardSpiDevice		@@if_ts_show_sd_pins
 		field = "SD logger rate",						sdCardLogFrequency
+		field = "SD logger mode",						sdTriggerLog
 
 	dialog = gpsReceiver, "GPS Receiver"
 		field = "gps RX",								gps_rx_pin
